refactor: シリアライザーと FirestoreStore のモジュールを整理
- SensorDataSerializer / JsonSerializer を storage → pipeline へ移動
  シリアライズはパイプラインの関心事のため
- FirestoreStore を storage から network へ移動し FirestoreSender にリネーム
  Firestore はリモートへの送信であり DataSender の実装として自然なため
- DataSender に onStop() デフォルトメソッドを追加
  FirestoreSender が未送信バッファをフラッシュするライフサイクルフックとして使用
- SensorPipeline.stop() で sender.onStop() を呼ぶよう更新
- storage/build.gradle.kts から Firebase 依存を削除
- network/build.gradle.kts に Firebase 依存を追加
- README / USAGE.md を実態に合わせて更新
