+++ network/src/main/java/com/example/wearos/network/DataSender.kt
@@ -2,4 +2,5 @@ package com.example.wearos.network
 
 interface DataSender {
     fun send(payload: String)
+    fun onStop() {}  // バッファを持つ実装がフラッシュ処理を行うためのライフサイクルフック
 }
