+++ pipeline/src/main/java/com/example/wearos/pipeline/SensorPipelineConfig.kt
@@ -2,8 +2,6 @@ package com.example.wearos.pipeline
 
 import com.example.wearos.network.DataSender
 import com.example.wearos.sensing.BaseSensorCollector
-import com.example.wearos.storage.JsonSerializer
-import com.example.wearos.storage.SensorDataSerializer
 import com.example.wearos.storage.SensorDataStore
 
 data class SensorPipelineConfig(
